requirements.txt
progressbar2==3.43.1
numpy==1.18.1
plotly==4.5.0
retrying==1.3.3
scipy==1.4.1
six==1.14.0
